+++ docs/changes.rst
@@ -15,6 +15,13 @@ To be released.  Alpha version.
 - Now :class:`~wand.image.Image` can take an optional ``format`` hint.
   [:issue:`32` by Michael Elovskikh]
 
+- Added :attr:`Image.alpha_channel <wand.image.Image.alpha_channel>`
+  property.  [:issue:`35` by Piotr Florczyk]
+
+- The default value of :meth:`Image.resize() <wand.image.Imageresize>`'s
+  ``filter`` option has changed from ``'triangle'`` to ``'undefined'``.
+  [:issue:`37` by Piotr Florczyk]
+
 - Added version data of the linked ImageMagick library into :mod:`wand.version`
   module: